pyproject.toml
[project]
name = "claude-claim-state-delta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.141.0",
    "google-genai>=1.14.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]
